Convert to annotation-based Spring context Fix: Bootstrap context only once, not per Gauge class initiated Fix: Move in named package bc. Spring can't handle classes without packages Update to current Spring 5.1.1 Update other dependencies
