--- src/main/resources/spring-module.xml
@@ -1,12 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans-2.5.xsd">
-
-    <bean id="paymentService" class="PaymentService">
-        <constructor-arg ref="debitCardProcessor"/>
-    </bean>
-
-    <bean id="debitCardProcessor" class="PaypalDebitCardProcessor"></bean>
-    <bean id="debitCard" class="DebitCard"></bean>
-</beans>
