+++ src/test/java/com/thoughtworks/gauge/example/java/DebitCardProcessor.java
@@ -1,3 +1,5 @@
+package com.thoughtworks.gauge.example.java;
+
 public interface DebitCardProcessor {
     Receipt charge(int amount);
 }
