+++ src/test/java/com/thoughtworks/gauge/example/java/Receipt.java
@@ -1,3 +1,5 @@
+package com.thoughtworks.gauge.example.java;
+
 public class Receipt {
 
     private String message;
