+++ src/test/java/com/thoughtworks/gauge/example/java/ReceiptStatus.java
@@ -1,3 +1,5 @@
+package com.thoughtworks.gauge.example.java;
+
 public enum ReceiptStatus {
     SUCCESS, FAILURE
 }
