+++ src/test/java/com/thoughtworks/gauge/example/java/SpringGaugeTestApp.java
@@ -1,0 +1,12 @@
+package com.thoughtworks.gauge.example.java;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Component scan finds all components/beans/... below this package.
+ */
+@Configuration
+@ComponentScan
+public class SpringGaugeTestApp {
+}
